requirements.txt
APScheduler==3.6.3
atomicwrites==1.4.0
attrs==20.3.0
beautifulsoup4==4.9.0
bs4==0.0.1
certifi==2019.11.28
cffi==1.14.3
chardet==3.0.4
colorama==0.4.4
cryptography==2.9.2
decorator==4.4.2
fire==0.3.1
idna==2.9
importlib-metadata==3.1.0
iniconfig==1.1.1
numpy==1.19.4
packaging==20.7
peewee==3.13.3
Pillow==7.1.2
pluggy==0.13.1
pony==0.7.13
ptbtest==1.3
py==1.9.0
pycparser==2.20
pyparsing==2.4.7
pytest==6.1.2
python-telegram-bot==13.0
pytz==2020.4
requests==2.23.0
six==1.14.0
soupsieve==2.0
termcolor==1.1.0
toml==0.10.2
tornado==6.1
tzlocal==2.1
urllib3==1.25.8
vk-api==11.8.0
zipp==3.4.0
